Move functions to go file where they are used
